+++ system/src/Grav/Common/Service/OutputServiceProvider.php
@@ -8,6 +8,8 @@
 
 namespace Grav\Common\Service;
 
+use Grav\Common\Page\Page;
+use Grav\Common\Twig\Twig;
 use Pimple\Container;
 use Pimple\ServiceProviderInterface;
 
@@ -16,7 +18,13 @@ class OutputServiceProvider implements ServiceProviderInterface
     public function register(Container $container)
     {
         $container['output'] = function ($c) {
-            return $c['twig']->processSite($c['page']->templateFormat());
+            /** @var Twig $twig */
+            $twig = $c['twig'];
+
+            /** @var Page $page */
+            $page = $c['page'];
+
+            return $twig->processSite($page->templateFormat());
         };
     }
 }