+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+# Feature/Objects Branch
+## mm/dd/2017
+
+1. [](#new)
+    * Added `Pages::baseUrl()`, `Pages::homeUrl()` and `Pages::url()` functions
+    * Added `Debugger::getCaller()` to figure out where the method was called from
+    * Added support for custom output providers like Slim Framework
+    * Added `Grav\Framework\Collection` classes for creating collections
+    * Added `Grav\Framework\ContentBlock` classes which add better support for nested HTML blocks with assets
+    * Added `Grav\Framework\Object` classes to support general objects and their collections
+    * Deprecated GravTrait
+1. [](#improved)
+    * Improve error handling in Folder::move()
+    * Added extra parameter for Twig::processSite() to include custom context
+
 # v1.2.5
 ## 04/xx/2017